requirements.txt
apache-airflow-providers-postgres>=5.0.0
psycopg2-binary
